WIP on the fingerprint UI. It can be opened, it shows items images, and you can pick up and dust the brush.

+++ Source/Persistence/PersistState.h
@@ -132,7 +132,7 @@ inline void PersistState::Xfer(const char* name, std::vector<T>& vector, bool lo
         {
             if(i < vector.size())
             {
-                Xfer("", vector[i]);
+                Xfer("", vector[static_cast<size_t>(i)]);
             }
             else
             {
